Added LinearHtml - a multi-band document model
Experimental use of separate arrays for content, attributes, formatting and meta items. Also experimenting with the use of format and element type indexes, allowing simpler data types (numbers) to be used in the actual data arrays, each a reference (by value) to an actual format object or element type string in the respective indexes.